style(http): Aggregate handlers for better usuability
